requirements.txt
#How to use it :
# pip install -r requirements.txt

# pandas
pandas==0.18.0

# scikit-learn
scikit-learn==0.17.1

# NLTK
nltk==3.2

# python-docx
python-docx==0.8.6

# requests
requests==2.9.1

# beautifulsoup4
beautifulsoup4==4.4.1

# numpy
numpy==1.11.0

# xgboost
xgboost==0.4

# matplotlib
matplotlib==1.5.1

# seaborn
seaborn==0.7.1